PID working
-modified serial print functions for signed integers
-modified PID settings parameters
-added PID init function to set initial params
-fixed analog reading in AVR
-fixed io_get_spindle declaration
-added pid initialization to code

+++ src/interface/settings.h
@@ -75,7 +75,7 @@ extern "C"
                 uint8_t laser_mode;
 #endif
 #if PID_CONTROLLERS > 0
-                int16_t pid_gain[PID_CONTROLLERS][3];
+                float pid_gain[PID_CONTROLLERS][3];
 #endif
         } settings_t;
